pyproject.toml
[project]
name = "ytdownloader"
version = "1.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "ffmpeg-python>=0.2.0",
    "pytubefix>=8.13.1",
]
